chore: combine implementations for attribute tracking
